+++ app/routes/aspects.tsx
@@ -1,12 +1,6 @@
-import { json, LoaderFunction } from '@remix-run/node';
+import { json } from '@remix-run/node';
 import type { ActionFunctionArgs } from '@remix-run/node';
-import {
-  useLoaderData,
-  Link,
-  Outlet,
-  Form,
-  useActionData,
-} from '@remix-run/react';
+import { Link, Outlet, Form, useActionData } from '@remix-run/react';
 import { styled } from 'styled-components';
 import PageHeader from '~/components/PageHeader';
 import { PageBreadcrumb } from '~/components/PageBreadcrumb';
@@ -33,47 +27,6 @@ export const handle = {
 };
 
 export default function Aspects() {
-  const actionData = useActionData();
-
-  const MINOR_ASPECTS = [
-    {
-      title: 'Quincunx',
-      imageSrc: '',
-      description:
-        'When two planets are 150 degrees apart, they are in quincunx. This aspect can create a sense of unease and discomfort, but it can also be a source of growth and learning.',
-    },
-    {
-      title: 'Semi-Sextile',
-      imageSrc: '',
-      description:
-        'When two planets are 30 degrees apart, they are in semi-sextile. This aspect is considered harmonious and can indicate opportunities for growth and development.',
-    },
-    {
-      title: 'Semi-Square',
-      imageSrc: '',
-      description:
-        'When two planets are 45 degrees apart, they are in semi-square. This aspect can create tension and challenges, but it can also be a source of motivation and growth.',
-    },
-    {
-      title: 'Quintile',
-      imageSrc: '',
-      description:
-        'When two planets are 72 degrees apart, they are in quintile. This aspect can indicate creativity and inspiration, and it can be a source of growth and development.',
-    },
-    {
-      title: 'Biquintile',
-      imageSrc: '',
-      description:
-        'When two planets are 144 degrees apart, they are in biquintile. This aspect can indicate creativity and inspiration, and it can be a source of growth and development.',
-    },
-    {
-      title: 'Septile',
-      imageSrc: '',
-      description:
-        'When two planets are 51.4286 degrees apart, they are in septile. This aspect can indicate creativity and inspiration, and it can be a source of growth and development.',
-    },
-  ];
-
   return (
     <>
       <PageHeader />
@@ -82,19 +35,12 @@ export default function Aspects() {
         <h1>Major Aspects</h1>
         <LgDetailCard items={MAJOR_ASPECTS} />
         <Outlet />
-        <h1>What are my aspects?</h1>
-        <MyForm />
-
-        <h2>Data from API</h2>
-        <pre>{JSON.stringify(actionData, null, 2)}</pre>
       </AspectsContainer>
     </>
   );
 }
 
 export function MyForm() {
-  const actionData = useActionData<typeof action>();
-
   return (
     <Form method='post'>
       <label htmlFor='full_name'>Full Name:</label>
@@ -156,9 +102,9 @@ export async function action({ request }: ActionFunctionArgs) {
   const formData = await request.formData();
 
   formData.append('api_key', apiKey as string);
-  for (let pair of formData.entries()) {
-    console.log(pair[0] + ', ' + pair[1]);
-  }
+  // for (let pair of formData.entries()) {
+  //   console.log(pair[0] + ', ' + pair[1]);
+  // }
 
   const fetchOptions = {
     method: 'POST',
